rendezvous::rendezvous

Struct Simulation

pub struct Simulation {

    pub t: f64,
    pub g: f64,
    pub softening: f64,
    pub dt: f64,
    pub dt_last_done: f64,
    pub steps_done: usize,
    pub var_cfg: Option<Vec<VariationalConfig>>,
    pub run_state: RunState,
    pub root_size: f64,
    pub root_x: usize,
    pub root_y: usize,
    pub root_z: usize,
    pub box_size: Vec3,
    pub box_size_max: f64,
    pub n_ghost_x: usize,
    pub n_ghost_y: usize,
    pub n_ghost_z: usize,
    pub gravity: Gravity,
    pub collision: Collision,
    pub collision_resolve: Option<CollisionResolver>,
    pub boundary: Boundary,
    pub track_energy_offset: bool,
    pub energy_offset: f64,
    pub exit_max_distance: Option<f64>,
    pub exit_min_distance: Option<f64>,
    pub exact_finish_time: bool,
    pub ignore_gravity_terms: IgnoreGravityTerms,
    pub test_particle_kind: TestParticleKind,
    pub particles: Particles,
    pub odes: Vec<OdeState>,
    pub gravity_cs: Vec<Vec3>,
    pub integrator: Integrator,
    pub heartbeat: Option<SimulationHook>,
    pub additional_forces: Option<SimulationHook>,
    pub pre_time_step_modifications: Option<SimulationHook>,
    pub post_time_step_modifications: Option<SimulationHook>,
    pub tree: Option<TreeKind>,
    pub tree_needs_update: bool,
    pub opening_angle: f64,

}

Fields
t: f64

Current simulation time. Initialized to zero by default.

g: f64

Gravitational constant. Initialized to 1.0 by default.

softening: f64

Gravitational softening. Initialized to 0.0 by default.

dt: f64

Time step. Initialized to 0.001 by default.

dt_last_done: f64

Last successful time step.

steps_done: usize

Number of steps done.

var_cfg: Option<Vec<VariationalConfig>>run_state: RunStateroot_size: f64root_x: usizeroot_y: usizeroot_z: usizebox_size: Vec3box_size_max: f64n_ghost_x: usizen_ghost_y: usizen_ghost_z: usizegravity: Gravitycollision: Collisioncollision_resolve: Option<CollisionResolver>boundary: Boundarytrack_energy_offset: boolenergy_offset: f64exit_max_distance: Option<f64>exit_min_distance: Option<f64>exact_finish_time: boolignore_gravity_terms: IgnoreGravityTerms

This variable determines if the gravity form the central object is included in the gravity calculation.

In general, the integrators will set this variable automatically and nothing needs to be changed by the user.

test_particle_kind: TestParticleKindparticles: Particlesodes: Vec<OdeState>gravity_cs: Vec<Vec3>integrator: Integratorheartbeat: Option<SimulationHook>

Executed at each timestep once. Use this to do extra output/work during a simulation.

additional_forces: Option<SimulationHook>pre_time_step_modifications: Option<SimulationHook>post_time_step_modifications: Option<SimulationHook>tree: Option<TreeKind>tree_needs_update: boolopening_angle: f64
Implementations

impl Simulation

pub fn init(integrator: Integrator) -> Self

pub fn configure_box(&mut self, root_size: f64, x: usize, y: usize, z: usize)

pub fn add(&mut self, p: Particle)

pub fn is_particle_in_box(&self, p: &Particle) -> bool

pub fn is_particle_inside_node(&self, node_id: NodeId) -> bool

pub fn pre_time_step(&mut self, _dt: f64)

pub fn post_time_step(&mut self, _dt: f64)

pub fn coefficient_of_restitution(&self, _p1: &Particle, _p2: &Particle) -> f64

pub fn run_heartbeat(&mut self) -> ControlFlow<ExitStatus, ()>

pub fn sigint_received(&self) -> bool

pub fn check_exit( &mut self, tmax: f64, last_full_dt: &mut f64, ) -> ControlFlow<ExitStatus, ()>

pub fn check_boundary(&mut self)

fn energy(&self) -> f64

fn should_remove(&self, pi: usize) -> bool

fn remove(&mut self, pi: usize, keep_sorted: bool)

pub fn add_variational_first_order(&mut self, is_test_particle: bool)

Add and initialize a set of first-order variational particles.

Each call to this method adds one VariationalConfig to the simulation and each variational config adds one variational particle for each real particle.

If is_test_particle is true, then only one variational particle will be added If is_test_particle is false, then variational particles will be added for all real particles

pub fn add_variational_second_order(&mut self)

pub fn update_tree(&mut self)

fn update_tree_node(&mut self, tree: &mut TreeKind, node_id: NodeId)

pub fn update_tree_gravity_data(&mut self)

pub fn update_accelerations(&mut self)

pub fn update_accelerations_for_variational_particles(&mut self)

pub fn rescale_variational_particles(&mut self)

pub fn search_collisions(&mut self)

pub fn synchronize(&mut self)

pub fn integrate(&mut self, t_end: f64) -> ExitStatus

pub fn step(&mut self)

pub fn pre_force_step(&mut self)

pub fn post_force_step(&mut self)