Crate rendezvous

Modules

boundary

collision

eos

gbs

The Gragg-Bulirsch-Stoer integrator (short BS for Bulirsch-Stoer) is an adaptive integrator which uses Richardson extrapolation and the modified midpoint method to obtain solutions to ordinary differential equations.

gravity

ias15

integrator

janus

leapfrog

This is the standard leap frog integrator. It is second order and symplectic. No configuration is available (the timestep is set in the simulation structure).

mercurius

ode

particle

rendezvous

reversible

Implementation of algorithms discussed in Hernandez & Dehnen (2024)

saba

sei

trace

tree

utils

whfast